packaging: Add npm archive of revpi-nodes to install on target system
The server is only compatible with a specific version of the
node-red-contrib-revpi-nodes. The nodes are installed via npm, which
would install the latest version over the Internet. However, this
version does not necessarily fit the installed server. The Debian
package will now install the appropriate node-red-contrib-revpi-nodes
archive on the target system.

Signed-off-by: Sven Sager <s.sager@kunbus.com>