feat(deb): Remove dependency to node-red-contrib-revpi-node
The server for the Node-RED nodes can be installed on a system without
Node-RED. In this case, this serves as a pure server that is used over
the network by a Node-RED instance. On the external Node-RED instance,
the appropriate nodes must be installed in the correct version. This is
controlled by the node-red-contrib-revpi-nodes package, not by this
one.

Signed-off-by: Sven Sager <s.sager@kunbus.com>