feat(deb): Install python package as application
The application is installed in its own folder and not in Python's
module folder. The entry point is usually stored in /usr/bin as a link.
However, since in the previous versions of the Bullseye package this
link was stored in /usr/sbin, we keep the behavior for Bullseye. Only
from Bookworm will this be changed.

Signed-off-by: Sven Sager <s.sager@kunbus.com>