fix(systemd): Let the daemon restart on-failure
The value "on-abort" only restarts the service if systemd receives a
signal from the service. If there is an error in the program, e.g. an
empty config.rsc, the service crashes with an exit code != 0. This is
not treated with the value "on-abort". The correct value is
"on-failure", which tries to restart the service every 10 seconds. If a
PiCtory configuration is created in the meantime, the service runs
again without errors.

Signed-off-by: Sven Sager <s.sager@kunbus.com>