fix: Allow supported data types only
Only data types supported by the outputs are allowed. Since the server
does not yet support a replace_io file, these are limited to bool and
int. The value sent via Node-RED is always of type str. If the else did
not throw an exception, an attempt would be made to write the string
into an output, which would cause the cycle function to crash.

The data types are queried via "is" on the real types, since a
`isinstance(True, int)` would return a `True`. This should be avoided.

Signed-off-by: Sven Sager <s.sager@kunbus.com>