feat: Reload RevPiModIO after driver reset and config change
When a new configuration is created (config.rsc), it is loaded by a
"driver-reset". All systems that use the process image
`/dev/piControl0` must respond to these changes to load the changes.

If a "driver-reset" is detected, it is checked whether the `config.rsc`
has been changed. If this is the case, the RevPiModIO module will be
reloaded.

Signed-off-by: Sven Sager <s.sager@kunbus.com>